+++ spec/operating_systems/netbsd/qemu_vm.spec.ts
@@ -1,0 +1,42 @@
+import {Vm} from '../../../src/operating_systems/netbsd/qemu_vm'
+import * as arch from '../../../src/architecture'
+import {host} from '../../../src/host'
+import * as os from '../../../src/operating_systems/kind'
+import {Accelerator} from '../../../src/vm'
+import '../../../src/operating_systems/netbsd/netbsd'
+
+describe('NetBSD QemuVm', () => {
+  let memory = '5G'
+  let cpuCount = 10
+
+  let osKind = os.Kind.for('netbsd')
+  let architecture = arch.Architecture.for(arch.Kind.x86_64, host, osKind)
+  let config = {
+    memory: memory,
+    cpuCount: cpuCount,
+    diskImage: '',
+    ssHostPort: 0,
+    cpu: '',
+    accelerator: Accelerator.tcg,
+    machineType: '',
+    uuid: '',
+    resourcesDiskImage: '',
+    userboot: '',
+    firmware: ''
+  }
+  let vm = new Vm('', '', architecture, config)
+
+  let getFlagValue = (flag: string) => vm.command[vm.command.indexOf(flag) + 1]
+  let actualMemory = () => getFlagValue('-m')
+  let actualSmp = () => getFlagValue('-smp')
+
+  describe('command', () => {
+    it('constucts a command with the correct memory configuration', () => {
+      expect(actualMemory()).toEqual(memory)
+    })
+
+    it('constucts a command with the correct SMP configuration', () => {
+      expect(actualSmp()).toEqual(cpuCount.toString())
+    })
+  })
+})